% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessibility.R
\name{accessibility}
\alias{accessibility}
\title{Access the Accessibility Indicators Dataset for GB}
\usage{
accessibility(
  origins = NULL,
  service = "employment",
  mode = "public_transport"
)
}
\arguments{
\item{origins}{Vector of origin LSOA/DZ codes. If NULL (default), all locations are returned.}

\item{service}{Character string specifying the desired service (e.g., "employment", "supermarket", "school"). Default is "employment".}

\item{mode}{Character string specifying the transportation mode. Can be "public_transport" (default) or "car".}
}
\value{
A data frame containing the data from the desired dataset.
}
\description{
This function checks for the accessibility indicators dataset in the data directory,
downloads it if necessary, and then loads the desired dataset based on the provided service and mode.
The function can also subset the data based on provided origin locations.
}
\examples{
\dontrun{
accessibility(service = 'employment', mode = "public_transport")
}
}
