% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_destinations.R
\name{aggregate_destinations}
\alias{aggregate_destinations}
\title{Aggregate Destinations by LSOA Geometries in the UK}
\usage{
aggregate_destinations(destinations = NULL, quiet = FALSE)
}
\arguments{
\item{destinations}{An SF object containing point geometries representing the destinations.}
}
\value{
A data frame containing the count of points at each LSOA location.
}
\description{
This function aggregates destinations represented on statistical geometries.
It uses 2011 Census LSOA geometries or equivalent sources from InfuseUK.
Currently, the function only handles destinations represented by SF points.
Thus, it returns the count of points at each location.
}
\examples{
\dontrun{
# Create the data frame with destinations
destinations <- data.frame(
  lon = c(-1.57, -1.99, -2.76, -1.11, -1.49, -2.56, -0.421, -1.47, -0.163, -1.99),
  lat = c(53.8, 51.4, 51.5, 53.5, 52.9, 51.5, 51.5, 52.5, 51.5, 51.4)
)

# Convert the data frame to an sf object
destinations <- sf::st_as_sf(destinations, coords = c("lon", "lat"), crs = 4326)

# Run function
aggregate_destinations(destinations)
}

}
\keyword{internal}
