% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_accessibility_data.R
\name{download_accessibility_data}
\alias{download_accessibility_data}
\title{Download and Extract Accessibility Indicators Dataset for GB}
\usage{
download_accessibility_data()
}
\value{
NULL Invisibly. The function is called for its side effects.
}
\description{
This function downloads the accessibility indicators dataset for Great Britain from Zenodo,
extracts the content, and then removes the ZIP file.
}
\examples{
download_extract_GB_data()
}
