% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_lsoa_geoms.R
\name{download_lsoa_geoms}
\alias{download_lsoa_geoms}
\title{Download Lower Layer Super Output Areas and Data Zones 2011 geographies}
\usage{
download_lsoa_geoms(
  force_update = FALSE,
  data_dir = system.file(package = "AccessUK"),
  quiet = FALSE
)
}
\arguments{
\item{force_update}{Logical, whether to re-download the data if it already exists.}

\item{data_dir}{The directory where the data should be stored. Default is "data" in the package's installation directory.}

\item{quiet}{Logical, suppresses messages when set to TRUE.}
}
\value{
Path to the extracted data directory invisibly.
}
\description{
This function downloads the 2011 Lower Layer Super Output Areas and Data Zones
geographies for all of the UK from InFuse. It also extracts the content, and then removes the ZIP file.
}
\examples{
\dontrun{
download_lsoa_geoms()
}
}
