% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_nearest_opportunity.R
\name{estimate_nearest_opportunity}
\alias{estimate_nearest_opportunity}
\title{Estimate the Nearest Opportunity}
\usage{
estimate_nearest_opportunity(
  travel_matrix = NULL,
  travel_cost = NULL,
  weights = NULL,
  additional_group = NULL
)
}
\arguments{
\item{travel_matrix}{A directory path to the travel matrix, which can be in CSV or Parquet format.
Default is \code{NULL}.}

\item{travel_cost}{The cost of travel, which will be used to determine the shortest time expressions.
It must be a valid column name from the travel matrix.}

\item{weights}{A dataframe containing weights, which will be written to the database.
It must contain an 'id' column and one or more weight columns.}

\item{additional_group}{Optional. A column name for additional grouping in the SQL query.
Default is \code{NULL}.}
}
\value{
A dataframe containing the nearest opportunities based on the given travel matrix,
travel cost, and weights.
}
\description{
This function estimates the nearest opportunity based on given travel matrix,
travel cost, and weights.
}
\examples{
\dontrun{
result <- estimate_nearest_opportunity(
  travel_matrix = "/path/to/travel_matrix",
  travel_cost = "cost_column",
  weights = data.frame(id = 1:3, weight1 = c(10, 20, 30))
)
}

}
