% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_tm_directory.R
\name{format_tm_directory}
\alias{format_tm_directory}
\title{Format Travel Matrix Directory}
\usage{
format_tm_directory(travel_matrix)
}
\arguments{
\item{travel_matrix}{A character string specifying the path to the directory
containing the travel matrix files.}
}
\value{
A character string with the formatted directory path, either pointing
to parquet or csv files.
}
\description{
Check for the presence of csv or parquet files in the specified directory and
format the directory path accordingly.
}
\examples{
\dontrun{
  formatted_path <- format_travel_matrix_directory("path/to/directory")
  print(formatted_path)
}


}
