% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_accessibility.R
\name{my_accessibility}
\alias{my_accessibility}
\title{Compute Accessibility Estimates}
\usage{
my_accessibility(destinations, time_cut, additional_group = NULL)
}
\arguments{
\item{destinations}{An SF object containing "POINT" geometries only or a data frame with weights aggregated by LSOA.
If the object is not SF, it is assumed to be weights aggregated by LSOA. This parameter is required.}

\item{time_cut}{A numeric vector specifying the time thresholds for the accessibility computation.
This parameter is required and cannot be empty.}

\item{additional_group}{Optional. A character string specifying an additional grouping variable.}
}
\value{
A data frame containing the computed accessibility estimates.
}
\description{
This function computes custom accessibility estimates using precomputed travel time matrices from the PTAI dataset and user-inputted destination data.
It can handle destinations represented by SF points or pre-aggregated weights by LSOA.
}
\details{
If \code{destinations} is an SF object, the function aggregates destinations by LSOA.
In the background, it uses 2011 geometries sourced from InFuse UK Data Service layers.
Please read the user Terms and Conditions.
}
\examples{
\dontrun{
# Example usage with SF points
destinations <- sf::st_as_sf(data.frame(
  lon = c(-1.57, -1.99, -2.76, -1.11, -1.49, -2.56, -0.421, -1.47, -0.163, -1.99),
  lat = c(53.8, 51.4, 51.5, 53.5, 52.9, 51.5, 51.5, 52.5, 51.5, 51.4)
), coords = c("lon", "lat"), crs = 4326)

# Run function
my_accessibility(
  destinations = destinations,
  time_cut = c(30, 45, 60)
)

# Example usage with aggregated data
weights_aggregated <- data.frame(
  id = c("E01000825", "E01000824", "E01000812"),
  n = c(10, 20, 30)
)

my_accessibility(
  destinations = weights_aggregated,
  time_cut = c(30, 45, 60)
)
}

}
