% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_nearest_opportunity.R
\name{my_nearest_opportunity}
\alias{my_nearest_opportunity}
\title{Compute Custom Nearest Opportunity Measures}
\usage{
my_nearest_opportunity(destinations, additional_group = NULL)
}
\arguments{
\item{destinations}{An SF object containing "POINT" geometries only or a data frame with weights aggregated by LSOA.
If the object is not SF, it is assumed to be weights aggregated by LSOA. This parameter is required.}

\item{additional_group}{Optional. A column name for additional grouping in the SQL query. Default is \code{NULL}.}
}
\value{
A data frame containing the nearest opportunities based on the PTAI travel matrix, travel cost, and weights.
}
\description{
This function computes custom nearest opportunity estimates using precomputed travel time matrices from the PTAI dataset and user-imputed destination data.
It can handle destinations represented by SF points or pre-aggregated weights by LSOA.
}
\details{
If \code{destinations} is an SF object, the function aggregates destinations by LSOA.
In the background, it uses 2011 geometries sourced from InFuse UK Data Service layers. Please read the user Terms and Conditions.
}
\examples{
\dontrun{
# Example usage with SF points
destinations <- sf::st_as_sf(data.frame(
  lon = c(-1.57, -1.99, -2.76, -1.11, -1.49),
  lat = c(53.8, 51.4, 51.5, 53.5, 52.9)
), coords = c("lon", "lat"), crs = 4326)

# Run function
nearest_opportunities <- my_nearest_opportunity(
  destinations = destinations
)

# Example usage with aggregated data
weights_aggregated <- data.frame(
  id = c("E01000825", "E01000824", "E01000812"),
  weight = c(10, 20, 30)
)

nearest_opportunities <- my_nearest_opportunity(
  destinations = weights_aggregated
)
}

}
