% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travel_time.R
\name{travel_time}
\alias{travel_time}
\title{Retrieve Travel Time Matrix Data}
\usage{
travel_time(from = NULL)
}
\arguments{
\item{from}{Vector of origin LSOA/DZ codes to get the travel time to all other. If NULL (default), all locations are returned.}
}
\value{
A data frame of the travel time matrix for the given origin.
}
\description{
Retrieves the travel time matrix (TTM) data by public transport based on a given origin.
}
\examples{
\dontrun{
travel_time(from = "S01006616")
}
}
