% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_accessibility_data.R
\name{download_accessibility_data}
\alias{download_accessibility_data}
\title{Download and extract Accessibility Indicators Dataset for GB}
\usage{
download_accessibility_data(
  force_update = FALSE,
  data_dir = system.file(package = "AccessUK"),
  quiet = FALSE
)
}
\arguments{
\item{force_update}{Logical, whether to re-download the data if it already exists.}

\item{data_dir}{The directory where the data should be stored. Default is "data" in the package's installation directory.}

\item{quiet}{Logical, suppresses messages when set to TRUE.}
}
\value{
Path to the extracted data directory invisibly.
}
\description{
This function downloads the accessibility indicators dataset for Great Britain from Zenodo,
extracts the content, and then removes the ZIP file.
}
\examples{
download_accessibility_data()
}
